% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkmeans.R
\name{kkmeans}
\alias{kkmeans}
\title{kkmeans}
\usage{
kkmeans(
  x,
  nCluster,
  nStart = 10,
  weights = NULL,
  kernel = "linear",
  kparam = 1,
  opt = TRUE,
  ...
)
}
\arguments{
\item{x}{input data}
}
\value{
out list of clusters
}
\description{
fitting kkmeans
}
