% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{skkm}
\alias{skkm}
\title{skkm}
\usage{
skkm(
  x,
  nCluster,
  nStart = 10,
  s = 1.5,
  weights = NULL,
  kernel = "linear",
  kparam = 1,
  search = "exact",
  opt = TRUE,
  ...
)
}
\arguments{
\item{x}{input data}
}
\value{
out list of clusters
}
\description{
fitting skkm
}
