% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkmeans.R
\name{tune.kkmeans}
\alias{tune.kkmeans}
\title{tune.skkm}
\usage{
tune.kkmeans(
  x,
  nCluster,
  nPerms = 20,
  nStart = 10,
  weights = NULL,
  kernel = "linear",
  kparam = 1,
  opt = TRUE,
  nCores = 1,
  ...
)
}
\arguments{
\item{x}{input data}
}
\value{
out results of kkmeans
}
\description{
tuning kkmeans parameter
}
