% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tune.skkm}
\alias{tune.skkm}
\title{tune.skkm}
\usage{
tune.skkm(
  x,
  nCluster,
  nPerms = 20,
  s = NULL,
  ns = 100,
  nStart = 10,
  weights = NULL,
  kernel = "linear",
  kparam = 1,
  opt = TRUE,
  nCores = 1,
  ...
)
}
\arguments{
\item{x}{input data}
}
\value{
s tuning parameter
}
\description{
tuning skkm parameter
}
